Added documentary comments and implemented TimeLine tracking Slider.
Created a method to bind width and height of the media view to the window size. Initialize method calls it instead of doing it all inside itself. Added comments before each method to make the code more readable. Implemented slider and time display label elements. The code for binding currentTime of mediaPlayer object is working although once called in Initialize method it doesn't work. It just has to be called once the mediaPlayer loaded the file and the duration is ready.